Make hover text trusted to enable clickable links
In order to be able to offer the user on hovering over a type error to
display a detailed explanation of that error, we need to be able to
display clickable links as part of hover messages. To that end this commit
adds a hook that sets the 'trusted' flag on the hover text.